+++ src/gga_x_cap.c
@@ -44,9 +44,9 @@ void
 XC(hyb_gga_xc_cap0_init)(XC(func_type) *p)
 {
   static int   funcs_id  [2] = {XC_GGA_X_CAP, XC_GGA_C_PBE};
-  static FLOAT funcs_coef[2] = {1.0, 1.0};
+  static double funcs_coef[2] = {1.0, 1.0};
   /* Can't init this above */
-  static const FLOAT a0 = 1.0/4.0;
+  static const double a0 = 1.0/4.0;
   funcs_coef[0]=1.0-a0;
   
   XC(mix_init)(p, 2, funcs_id, funcs_coef);
